Combine two video util files
